+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -70,6 +70,7 @@ public class ProjectManagerConst {
     public final static String SET_NOTIFICATION_AS_READ_ACTION = "SET_NOTIFICATION_AS_READ";
     public final static String FETCH_PROJECT_ACTION = "FETCH_PROJECT_ACTION";
     public final static String FETCH_PROJECT_STATES_ACTION = "FETCH_PROJECT_STATES";
+    public final static String FETCH_ALL_REGISTERED_BRIDGEHEADS_ACTION = "FETCH_ALL_REGISTERED_BRIDGEHEADS";
 
 
     // REST Services
@@ -128,6 +129,7 @@ public class ProjectManagerConst {
     public final static String FETCH_OTHER_DOCUMENTS = "/other-documents";
     public final static String FETCH_NOTIFICATIONS = "/notifications";
     public final static String SET_NOTIFICATION_AS_READ = "/read-notification";
+    public final static String FETCH_ALL_REGISTERED_BRIDGEHEADS = "/bridgeheads";
 
     // REST Parameters
     public final static String PROJECT_CODE = "project-code";