+++ CHANGELOG.md
@@ -84,3 +84,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch User Visible Projects and User Visible Notifications without constraints
 - Fetch Project
 - Fetch Project States
+- Fetch registered bridgeheads for project manager admin